pyproject.toml
[tool.poetry]
name = "opentelemetry-python-examples"
version = "0.1.0"
description = ""
authors = ["avpps <avppsdev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opentelemetry-sdk = "^1.21.0"
opentelemetry-api = "^1.21.0"
opentelemetry-exporter-otlp = "^1.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
